requirements.txt
numpy~=1.26.3
Pillow~=10.2.0
torch>=1.7
torchvision>=0.8.0

aiogram~=3.3.0
environs~=10.3.0